Change location of tests
